always perform `/auth` check on callback_url(s)
